Add `checkGooseEgg` utility method.
